Added CRUD operations for drivers place and runs
Enhanced the functionality of the drivers place and runs' endpoints to support all CRUD operations. In preparation for these changes, the ID field has been added to the DriverPlace model. The update_data and delete_data methods have also been added to _DBConnection in database/cars.py. The host details in sql_credentials.json and the requirements.txt were updated as well.