Adjust default value for weight in Front Interaction model
The weight attribute's default value in the Front Interaction model has been adjusted. Priorly, this was set to a direct 0, however, it has now been changed to a Decimal 0 in order to maintain data type consistency and prevent potential errors in the future. This change clarifies and simplifies future code interpretations and operations involving this attribute.